Replace Travis CI with Github Actions
The motiviation behind this change is that Github Actions is obviously better integrated in Github than Travis CI and that it will be easy to run test against non Linux platforms, e.g. macOS or Windows.
Tests are now run for the three most recent Go versions, i.e. Go 1.13, 1.14 and 1.15 .
A new 'test' target was added to the Makefile.